3.10app文件夹

+++ app/src/main/java/com/linhuiba/business/model/CommentScoreModel.java
@@ -5,94 +5,94 @@ package com.linhuiba.business.model;
  */
 
 public class CommentScoreModel {
-    private Integer composite_score;
-    private Integer count_of_composite_score;
-    private Integer score_of_visitorsflowrate;
-    private Integer count_of_visitorsflowrate_reviews;
-    private Integer score_of_userparticipation;
-    private Integer count_of_userparticipation_reviews;
-    private Integer score_of_propertymatching;
-    private Integer count_of_propertymatching_reviews;
-    private Integer score_of_goalcompletion;
-    private Integer count_of_goalcompletion_reviews;
+    private Double composite_score;
+    private Double count_of_composite_score;
+    private Double score_of_visitorsflowrate;
+    private Double count_of_visitorsflowrate_reviews;
+    private Double score_of_userparticipation;
+    private Double count_of_userparticipation_reviews;
+    private Double score_of_propertymatching;
+    private Double count_of_propertymatching_reviews;
+    private Double score_of_goalcompletion;
+    private Double count_of_goalcompletion_reviews;
 
-    public Integer getComposite_score() {
+    public Double getComposite_score() {
         return composite_score;
     }
 
-    public void setComposite_score(Integer composite_score) {
+    public void setComposite_score(Double composite_score) {
         this.composite_score = composite_score;
     }
 
-    public Integer getCount_of_composite_score() {
+    public Double getCount_of_composite_score() {
         return count_of_composite_score;
     }
 
-    public void setCount_of_composite_score(Integer count_of_composite_score) {
+    public void setCount_of_composite_score(Double count_of_composite_score) {
         this.count_of_composite_score = count_of_composite_score;
     }
 
-    public Integer getScore_of_visitorsflowrate() {
+    public Double getScore_of_visitorsflowrate() {
         return score_of_visitorsflowrate;
     }
 
-    public void setScore_of_visitorsflowrate(Integer score_of_visitorsflowrate) {
+    public void setScore_of_visitorsflowrate(Double score_of_visitorsflowrate) {
         this.score_of_visitorsflowrate = score_of_visitorsflowrate;
     }
 
-    public Integer getCount_of_visitorsflowrate_reviews() {
+    public Double getCount_of_visitorsflowrate_reviews() {
         return count_of_visitorsflowrate_reviews;
     }
 
-    public void setCount_of_visitorsflowrate_reviews(Integer count_of_visitorsflowrate_reviews) {
+    public void setCount_of_visitorsflowrate_reviews(Double count_of_visitorsflowrate_reviews) {
         this.count_of_visitorsflowrate_reviews = count_of_visitorsflowrate_reviews;
     }
 
-    public Integer getScore_of_userparticipation() {
+    public Double getScore_of_userparticipation() {
         return score_of_userparticipation;
     }
 
-    public void setScore_of_userparticipation(Integer score_of_userparticipation) {
+    public void setScore_of_userparticipation(Double score_of_userparticipation) {
         this.score_of_userparticipation = score_of_userparticipation;
     }
 
-    public Integer getCount_of_userparticipation_reviews() {
+    public Double getCount_of_userparticipation_reviews() {
         return count_of_userparticipation_reviews;
     }
 
-    public void setCount_of_userparticipation_reviews(Integer count_of_userparticipation_reviews) {
+    public void setCount_of_userparticipation_reviews(Double count_of_userparticipation_reviews) {
         this.count_of_userparticipation_reviews = count_of_userparticipation_reviews;
     }
 
-    public Integer getScore_of_propertymatching() {
+    public Double getScore_of_propertymatching() {
         return score_of_propertymatching;
     }
 
-    public void setScore_of_propertymatching(Integer score_of_propertymatching) {
+    public void setScore_of_propertymatching(Double score_of_propertymatching) {
         this.score_of_propertymatching = score_of_propertymatching;
     }
 
-    public Integer getCount_of_propertymatching_reviews() {
+    public Double getCount_of_propertymatching_reviews() {
         return count_of_propertymatching_reviews;
     }
 
-    public void setCount_of_propertymatching_reviews(Integer count_of_propertymatching_reviews) {
+    public void setCount_of_propertymatching_reviews(Double count_of_propertymatching_reviews) {
         this.count_of_propertymatching_reviews = count_of_propertymatching_reviews;
     }
 
-    public Integer getScore_of_goalcompletion() {
+    public Double getScore_of_goalcompletion() {
         return score_of_goalcompletion;
     }
 
-    public void setScore_of_goalcompletion(Integer score_of_goalcompletion) {
+    public void setScore_of_goalcompletion(Double score_of_goalcompletion) {
         this.score_of_goalcompletion = score_of_goalcompletion;
     }
 
-    public Integer getCount_of_goalcompletion_reviews() {
+    public Double getCount_of_goalcompletion_reviews() {
         return count_of_goalcompletion_reviews;
     }
 
-    public void setCount_of_goalcompletion_reviews(Integer count_of_goalcompletion_reviews) {
+    public void setCount_of_goalcompletion_reviews(Double count_of_goalcompletion_reviews) {
         this.count_of_goalcompletion_reviews = count_of_goalcompletion_reviews;
     }
 }
