+++ app/src/main/java/com/linhuiba/business/model/PhysicalResourceModel.java
@@ -52,7 +52,7 @@ public class PhysicalResourceModel implements Serializable {
     private ArrayList<CaseListModel> activity_cases;
     //广告
     private Field_AddResourceCreateItemModel ad_type;
-    private Integer average_score;
+    private Double average_score;
     //拼团 //或场地
     private Field_AddResourceCreateItemModel field_type;
     private Field_AddResourceCreateItemModel activity_type;
@@ -390,11 +390,11 @@ public class PhysicalResourceModel implements Serializable {
         this.ad_type = ad_type;
     }
 
-    public Integer getAverage_score() {
+    public Double getAverage_score() {
         return average_score;
     }
 
-    public void setAverage_score(Integer average_score) {
+    public void setAverage_score(Double average_score) {
         this.average_score = average_score;
     }
